+++ packages/patternfly-4/gatsby-theme-patternfly-org/gatsby-node.js
@@ -7,13 +7,13 @@ const { slugger } = require('./helpers/slugger');
 
 // Add map PR-related environment variables to GraphQL
 exports.sourceNodes = ({ actions, createNodeId, createContentDigest }) => {
-  const num = process.env.CIRCLE_PR_NUMBER || process.env.PR_NUMBER;
-  const url = process.env.CIRCLE_PULL_REQUEST;
+  const num = process.env.CIRCLE_PR_NUMBER || process.env.PR_NUMBER || '';
+  const url = process.env.CIRCLE_PULL_REQUEST || '';
   // Docs https://www.gatsbyjs.org/docs/actions/#createNode
   actions.createNode({
     name: 'PR_INFO',
-    num: num || '',
-    url: url || '',
+    num,
+    url,
     id: createNodeId('PR_INFO'),
     parent: null,
     children: [],
@@ -114,7 +114,7 @@ exports.onCreateNode = ({ node, actions, getNode }) => {
       addedFileFieldsToSchema = true;
     }
     if (node.extension === 'hbs') {
-      // Partial name has always come from file name in patternfly-next
+      // Partial name has always come from file name in patternfly
       createNodeField({
         node,
         name: 'name',
@@ -199,7 +199,7 @@ exports.createPages = ({ actions, graphql }, pluginOptions) => graphql(`
         if (source === 'react') {
           sourceLink += 'patternfly-react/blob/master/';
         } else if (source === 'core') {
-          sourceLink += 'patternfly-next/blob/master/';
+          sourceLink += 'patternfly/blob/master/';
         } else if (source === 'shared') {
           sourceLink += 'patternfly-org/blob/master/packages/patternfly-4/gatsby-theme-patternfly-org/';
         } else {
@@ -235,7 +235,7 @@ exports.createPages = ({ actions, graphql }, pluginOptions) => graphql(`
             title,
             // For top of TOC, dynamic classNames in Example.js, and some feature flags
             source,
-            // To render static example HTML from patternfly-next
+            // To render static example HTML from patternfly
             htmlExamples: source === 'core' ? examples : undefined,
             // To hide the banner for core/React sites
             showBanner,